backend/requirements.txt
# FastAPI
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.27
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Pydantic
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# JWT & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.1.2

# OAuth2
authlib==1.3.0

# Celery & Redis
celery==5.3.6
redis==5.0.1

# HTTP Client
httpx==0.26.0

# S3/MinIO
boto3==1.34.37
minio==7.2.3

# Google Speech-to-Text (для голосовых сообщений)
google-cloud-speech==2.24.0
SpeechRecognition==3.10.1
pydub==0.25.1

# PDF Generation
reportlab==4.0.9
weasyprint==60.2

# Utils
python-dotenv==1.0.1
pytz==2024.1

